# src/main/resources/static/openapi/exercise.yaml
openapi: 3.0.3
info:
  title: Exercise APIs
  version: 1.0.0

paths:
  /v1/authentication/users:
    post:
      summary: Authenticate student and teacher and return a JWT token
      operationId: getUserToken
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetUserTokenRequest'
      responses:
        '200':
          description: JWT token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTokenResponse'

  /v1/authentication/admins:
    post:
      summary: Authenticate admin and return a JWT token
      operationId: getAdminToken
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetAdminTokenRequest'
      responses:
        '200':
          description: JWT token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTokenResponse'

  /v1/student/teachers:
    get:
      summary: List all teachers
      operationId: listTeachersForStudent
      tags:
        - Student
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of teachers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeachersForStudentResponse'

  /v1/student/teachers/{teacherId}/availabilities:
    get:
      summary: Fetch teacher availability for a specific week
      operationId: fetchTeacherAvailabilities
      tags:
        - Student
      security:
        - bearerAuth: []
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
        - name: isoWeek
          in: query
          required: true
          schema:
            type: string
            example: '2025-W20'
      responses:
        '200':
          description: Teacher's availability for the given week
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherAvailabilitiesResponse'

  /v1/student/schedules:
    post:
      summary: Schedule a new class
      operationId: scheduleClass
      tags:
        - Student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleClassRequest'
      responses:
        '201':
          description: Class scheduled
    
  /v1/student/schedules/upcoming:
    get:
      summary: List upcoming scheduled classes
      operationId: listUpcomingScheduledClasses
      tags:
        - Student
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of upcoming schedules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledClassesResponse'

  /v1/student/schedules/past:
    get:
      summary: List past scheduled classes
      operationId: listPastScheduledClasses
      tags:
        - Student
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of past schedules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledClassesResponse'

  /v1/student/schedules/{scheduleId}:
    delete:
      summary: Cancel a scheduled class
      operationId: cancelClass
      tags:
        - Student
      security:
        - bearerAuth: []
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Class deleted

  /v1/student/points:
    get:
      summary: Get student points
      operationId: getStudentPoints
      tags:
        - Student
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentPointsResponse'

  /v1/student/points/records:
    get:
      summary: List student points records
      operationId: listStudentPointsRecords
      tags:
        - Student
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of student points records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentPointsRecordsResponse'

  /v1/teacher/availabilities/recurring:
    post:
      summary: Add a recurring availability slot
      operationId: addTeacherRecurringAvailability
      tags:
        - Teacher
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTeacherRecurringAvailabilityRequest'
      responses:
        '201':
          description: Slot added
    get:
      summary: List upcoming recurring availability slots
      operationId: listTeacherRecurringAvailabilities
      tags:
        - Teacher
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of recurring time slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherRecurringAvailabilitiesResponse'

  /v1/teacher/availabilities/recurring/{recurringId}:
    delete:
      summary: Delete a recurring availability slot
      operationId: deleteTeacherRecurringAvailability
      tags:
        - Teacher
      security:
        - bearerAuth: []
      parameters:
        - name: recurringId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Slot deleted

  /v1/teacher/availabilities/exception:
    post:
      summary: Add an availability exception
      operationId: addTeacherExceptionAvailability
      tags:
        - Teacher
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTeacherExceptionAvailabilityRequest'
      responses:
        '201':
          description: Exception added
    get:
      summary: List upcoming availability exceptions
      operationId: listUpcomingTeacherExceptionAvailabilities
      tags:
        - Teacher
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of exceptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherExceptionAvailabilitiesResponse'

  /v1/teacher/availabilities/exception/{exceptionId}:
    delete:
      summary: Delete an availability exception
      operationId: deleteTeacherExceptionAvailability
      tags:
        - Teacher
      security:
        - bearerAuth: []
      parameters:
        - name: exceptionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Exception deleted

  /v1/teacher/availabilities/unavailable:
    post:
      summary: Add an unavailable date
      operationId: addTeacherUnavailableDate
      tags:
        - Teacher
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTeacherUnavailableDateRequest'
      responses:
        '201':
          description: Date added
    get:
      summary: List upcoming teacher's unavailable dates
      operationId: listUpcomingTeacherUnavailableDates
      tags:
        - Teacher
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of unavailable dates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherUnavailableDatesResponse'

  /v1/teacher/availabilities/unavailable/{unavailableDateId}:
    delete:
      summary: Delete an unavailable date
      operationId: deleteTeacherUnavailableDate
      tags:
        - Teacher
      security:
        - bearerAuth: []
      parameters:
        - name: unavailableDateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Date deleted

  /v1/admin/students:
    post:
      summary: Create a new student
      operationId: createStudent
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
      responses:
        '201':
          description: Student created

  /v1/admin/students/{studentPhone}:
    get:
      summary: Get student by phone
      operationId: getStudentByPhone
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: studentPhone
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'

  /v1/admin/students/{studentId}:
    delete:
      summary: Delete a student
      operationId: deleteStudent
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Student deleted

  /v1/admin/teachers:
    post:
      summary: Create a new teacher
      operationId: createTeacher
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherRequest'
      responses:
        '201':
          description: Teacher created
    get:
      summary: List all teachers
      operationId: listTeachersForAdmin
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of teachers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeachersForAdminResponse'
                  

  /v1/admin/teachers/{teacherId}:
    delete:
      summary: Delete a teacher
      operationId: deleteTeacher
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Teacher deleted

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    GetUserTokenRequest:
      type: object
      required:
        - phone
        - password
      properties:
        phone:
          type: string
        password:
          type: string

    GetAdminTokenRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    UserTokenResponse:
      type: object
      properties:
        role:
          type: string
          enum: [STUDENT, TEACHER]
        token:
          type: string

    AdminTokenResponse:
      type: object
      properties:
        token:
          type: string

    StudentResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time

    StudentPointsResponse:
      type: object
      properties:
        points:
          type: integer

    StudentPointsRecordsResponse:
      type: object
      properties:
        pointsRecords:
          type: array
          items:
            $ref: '#/components/schemas/StudentPointsRecordResponse'
    
    StudentPointsRecordResponse:
      type: object
      properties:
        id:
          type: integer
        adjustedPoints:
          type: integer
        reason:
          type: string
        pointsAfterAdjustment:
          type: integer

    CreateStudentRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        points:
          type: integer
          default: 0

    TeachersForStudentResponse:
      type: object
      properties:
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/TeacherForStudentResponse'

    TeacherForStudentResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

    TeachersForAdminResponse:
      type: object
      properties:
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/TeacherForAdminResponse'

    TeacherForAdminResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateTeacherRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string

    TeacherAvailabilitiesResponse:
      type: object
      properties:
        teacherAvailabilities:
          type: array
          items:
            $ref: '#/components/schemas/TeacherAvailabilityResponse'

    TeacherAvailabilityResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
        timeslots:
          type: array
          items:
            type: string
            format: time

    AddTeacherRecurringAvailabilityRequest:
      type: object
      properties:
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time

    TeacherRecurringAvailabilitiesResponse:
      type: object
      properties:
        recurringAvailabilities:
          type: array
          items:
            $ref: '#/components/schemas/TeacherRecurringAvailabilityResponse'

    TeacherRecurringAvailabilityResponse:
      type: object
      properties:
        id:
          type: integer
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time

    AddTeacherExceptionAvailabilityRequest:
      type: object
      properties:
        exceptionDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time

    TeacherExceptionAvailabilitiesResponse:
      type: object
      properties:
        exceptionAvailabilities:
          type: array
          items:
            $ref: '#/components/schemas/TeacherExceptionAvailabilityResponse'

    TeacherExceptionAvailabilityResponse:
      type: object
      properties:
        id:
          type: integer
        exceptionDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string

    AddTeacherUnavailableDateRequest:
      type: object
      properties:
        unavailableDate:
          type: string
          format: date

    TeacherUnavailableDatesResponse:
      type: object
      properties:
        unavailableDates:
          type: array
          items:
            $ref: '#/components/schemas/TeacherUnavailableDateResponse'

    TeacherUnavailableDateResponse:
      type: object
      properties:
        id:
          type: integer
        unavailableDate:
          type: string
          format: date

    ScheduleClassRequest:
      type: object
      properties:
        teacherId:
          type: integer
        classDate:
          type: string
          format: date
        startTime:
          type: string
          format: time

    ScheduledClassesResponse:
      type: object
      properties:
        scheduledClasses:
          type: array
          items:
            $ref: '#/components/schemas/ScheduledClassResponse'

    ScheduledClassResponse:
      type: object
      properties:
        id:
          type: integer
        teacherName:
          type: string
        classDate:
          type: string
          format: date
        startTime:
          type: string
          format: time